Use getDouble with SQL function AVG()
MySQL that restricts float being queried with getLong caused an
SQLException

All usages of getLong with AVG function replaced with
getDouble cast to long.

Affects issues:
- Fixed #1181

+++ Plan/common/src/main/java/com/djrapitops/plan/storage/database/queries/analysis/ActivityIndexQueries.java
@@ -292,7 +292,7 @@ public class ActivityIndexQueries {
 
                 @Override
                 public Long processResults(ResultSet set) throws SQLException {
-                    return set.next() ? set.getLong("average") : 0;
+                    return set.next() ? (long) set.getDouble("average") : 0;
                 }
             });
         };
@@ -326,7 +326,7 @@ public class ActivityIndexQueries {
 
                 @Override
                 public Long processResults(ResultSet set) throws SQLException {
-                    return set.next() ? set.getLong("average") : 0;
+                    return set.next() ? (long) set.getDouble("average") : 0;
                 }
             });
         };
@@ -361,7 +361,7 @@ public class ActivityIndexQueries {
 
                 @Override
                 public Long processResults(ResultSet set) throws SQLException {
-                    return set.next() ? set.getLong("average") : 0;
+                    return set.next() ? (long) set.getDouble("average") : 0;
                 }
             });
         };
