+++ Plan/common/src/main/java/com/djrapitops/plan/storage/database/queries/analysis/NetworkActivityIndexQueries.java
@@ -282,7 +282,7 @@ public class NetworkActivityIndexQueries {
 
                 @Override
                 public Long processResults(ResultSet set) throws SQLException {
-                    return set.next() ? set.getLong("average") : 0;
+                    return set.next() ? (long) set.getDouble("average") : 0;
                 }
             });
         };
@@ -314,7 +314,7 @@ public class NetworkActivityIndexQueries {
 
                 @Override
                 public Long processResults(ResultSet set) throws SQLException {
-                    return set.next() ? set.getLong("average") : 0;
+                    return set.next() ? (long) set.getDouble("average") : 0;
                 }
             });
         };
@@ -347,7 +347,7 @@ public class NetworkActivityIndexQueries {
 
                 @Override
                 public Long processResults(ResultSet set) throws SQLException {
-                    return set.next() ? set.getLong("average") : 0;
+                    return set.next() ? (long) set.getDouble("average") : 0;
                 }
             });
         };
